requirements.txt
python==3.13.0
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
django-filter==24.3
idna==3.10
jwt==1.3.1
kombu==5.4.2
pip==25.0
prompt_toolkit==3.0.50
psycopg==3.2.4
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13